+++ frontend/src/screens/DriverSignupScreen.js
@@ -1,5 +1,6 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useCallback } from "react";
 import { View, Text, TextInput, Button, TouchableOpacity, ActivityIndicator, Image, Alert, StyleSheet, ScrollView } from "react-native";
+import { useNavigation, useFocusEffect } from "@react-navigation/native";
 import * as DocumentPicker from "expo-document-picker";
 import * as FileSystem from "expo-file-system";
 import i18n from '../i18n/i18n.js';
@@ -8,9 +9,10 @@ import Toast from 'react-native-toast-message';
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { jwtDecode } from "jwt-decode";
 
-const DriverSignupScreen = ({ navigation }) => {
+const DriverSignupScreen = () => {
   const [showVerification, setShowVerification] = useState(false)
   const [verificationCode, setVerificationCode] = useState("");
+  const navigation = useNavigation();
 
   
   // State to store form input values
@@ -153,7 +155,7 @@ const DriverSignupScreen = ({ navigation }) => {
     props: { showIcon: true }, // Custom Prop for Icon
   });
   setShowVerification(true)
-  // navigation.navigate("DriverSignin");
+  // navigation.replace("DriverSignin");
      } catch (error) {
       console.log(error)
   setLoading(false);
@@ -188,34 +190,51 @@ const DriverSignupScreen = ({ navigation }) => {
       await AsyncStorage.setItem("userType", "driver");
       await AsyncStorage.setItem("userData", JSON.stringify(response.driver));
       setLoading(false)
-      navigation.navigate("DriverHome"); // Redirect to home
+      navigation.replace("DriverHome"); // Redirect to home
     } catch (error) {
       Toast.show({ type: "error", text1: "Verification Failed", text2: error.message });
       setLoading(false)
     }
   };
 
-  useEffect(() => {
-    const checkDriverToken = async () => {
+  const initializeDriverSignUpScreen = async () => {
+    try {
       const token = await AsyncStorage.getItem("userToken");
-      if (token) {
+      const userType = await AsyncStorage.getItem("userType");
+      console.log("Home.js token, userType,:", token, userType);
+      console.log("Home.js type of token:", typeof token);
+
+      if (token && userType) {
+        console.log("Home.js token, userType,:", token, userType);
         try {
           const decoded = jwtDecode(token);
-          console.log("DriverSignupScreen.js checkDriverToken decoded", decoded)
-          const now = Date.now() / 1000;
-          if (decoded.exp > now) {
-            navigation.replace("DriverHome");
+          const now = Date.now() / 1000; // in seconds
+          console.log("Home.js decoded:", decoded);
+          console.log("Home.js now:", now);
+          console.log("Home.js decoded.exp:", decoded.exp);
+          if (decoded.exp < now) {
+            await AsyncStorage.clear();
           } else {
-            await AsyncStorage.clear(); // Clear all AsyncStorage data
+            navigation.replace(userType === "driver" ? "DriverHome" : "ClientHome");
           }
-        } catch (error) {
-          console.error("Invalid token:", error);
-          await AsyncStorage.clear(); // Clear all AsyncStorage data
+        } catch (e) {
+          console.log("Invalid token");
+          console.log(e);
+          await AsyncStorage.clear();
         }
+      } else {
+        console.log("No token or userType found, staying on Home screen");
       }
-    };
-    checkDriverToken();
-  }, []);
+    } catch (error) {
+      console.error("Error during initialization:", error);
+    }
+  };
+
+  useFocusEffect(
+    useCallback(() => {
+      initializeDriverSignUpScreen();
+    }, [])
+  );
 
   return (
     <View style={styles.container}>
